move directory recipe
