+++ app/src/main/java/com/jin/honey/feature/recipe/ui/RecipeViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.jin.feature.ui.state.UiState
 import com.jin.domain.usecase.GetRecipeUseCase
-import com.jin.domain.model.recipe.RecipePreview
+import com.jin.domain.recipe.model.RecipePreview
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
