+++ app/src/main/java/com/jin/honey/feature/recipe/ui/content/RecipeContent.kt
@@ -9,7 +9,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import com.jin.domain.model.recipe.RecipeStep
+import com.jin.domain.recipe.model.RecipeStep
 
 @Composable
 fun RecipeContent(recipeSteps: List<RecipeStep>) {
