+++ core/database/src/main/java/com/jin/database/Converters.kt
@@ -6,7 +6,7 @@ import com.jin.domain.address.model.Address
 import com.jin.domain.cart.model.Cart
 import com.jin.domain.cart.model.IngredientCart
 import com.jin.domain.food.model.Ingredient
-import com.jin.domain.model.recipe.RecipeStep
+import com.jin.domain.recipe.model.RecipeStep
 import com.jin.network.NetworkProvider
 
 class Converters {
