+++ domain/src/main/java/com/jin/domain/usecase/SyncRecipesUseCase.kt
@@ -1,6 +1,6 @@
 package com.jin.domain.usecase
 
-import com.jin.domain.repositories.RecipeRepository
+import com.jin.domain.recipe.RecipeRepository
 
 class SyncRecipesUseCase(private val recipeRepository: RecipeRepository) {
     suspend operator fun invoke() {
